requirements.txt
bottle==0.12.19
bottle-websocket==0.2.9
cycler==0.10.0
Eel==0.14.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
pyFirmata==1.1.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.2
six==1.16.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
